se refactorizan los paquetes y las clases
