+++ app/src/main/java/com/jadevelopers/eden/EdenApplication.kt
@@ -1,0 +1,7 @@
+package com.jadevelopers.eden
+
+import android.app.Application
+
+class EdenApplication : Application() {
+
+}
