+++ app/src/main/java/com/jadevelopers/eden/database/ShoppingCarDb.kt
@@ -1,8 +1,8 @@
-package com.jadevelopers.eden
+package com.jadevelopers.eden.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.jadevelopers.eden.model.ShoppingCar
+import com.jadevelopers.eden.database.entities.ShoppingCar
 
 @Database(entities = [ShoppingCar::class], version = 1)
