+++ app/src/main/java/com/jadevelopers/eden/features/productdescription/DialogFragment.kt
@@ -1,4 +1,4 @@
-package com.jadevelopers.eden.view
+package com.jadevelopers.eden.features.productdescription
 
 import android.os.Bundle
 import android.view.LayoutInflater
