+++ app/src/main/res/layout/fragment_dialogs.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.DialogFragment">
+    tools:context=".features.productdescription.DialogFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintCantidad"
