+++ app/src/main/res/layout/fragment_shopping_car.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/padding"
-    tools:context=".view.ShoppingCarFragment">
+    tools:context=".features.shoppingcar.ShoppingCarFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/productsShoppingContainer"
