+++ app/src/main/res/layout/fragment_tracking.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.TrackingFragment"
+    tools:context=".features.tracking.TrackingFragment"
     android:background="@color/white">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
